refactoring the Content component
