+++ frontend/src/components/Content.tsx
@@ -2,51 +2,21 @@ import "regenerator-runtime/runtime";
 import {
   Dispatch,
   SetStateAction,
-  useEffect,
   useRef,
   useState,
   useMemo,
   useLayoutEffect,
 } from "react";
-import LatexDisplayer from "./Latex";
-import SpeechRecognition, {
-  useSpeechRecognition,
-} from "react-speech-recognition";
 import { motion } from "framer-motion";
-import { Textarea } from "./ui/textarea";
-import { Button } from "./ui/button";
-import {
-  Card,
-  CardContent,
-  CardHeader,
-  CardTitle,
-  CardDescription,
-} from "./ui/card";
-import { getCookie } from "cookies-next";
-
-import { Tabs, TabsContent, TabsList, TabsTrigger } from "./ui/tabs";
-import {
-  Mic,
-  MicOff,
-  Copy,
-  BookOpen,
-  Download,
-  Pencil,
-  Type,
-} from "lucide-react";
-import { toast } from "react-hot-toast";
-import RecordingManager, { TranscriptionMethod } from "../utils/recording";
+import { TranscriptionMethod } from "../utils/recording";
 import { TranscriptionToggle } from "./TranscriptionToggle";
 import { copyToClipboard, downloadLatexAsPDF, debounce } from "@/lib/utils";
-import {
-  DefaultSpinner,
-  Tldraw,
-  createTLStore,
-  getSnapshot,
-  loadSnapshot,
-  exportToBlob,
-} from "tldraw";
+import { createTLStore } from "tldraw";
 import "tldraw/tldraw.css";
+import { handleReceiveDrawing, createDrawingEmitter } from "@/utils/drawing";
+import { useVoiceRecording } from "@/hooks/useVoiceRecording";
+import { EditorView } from "./room/EditorView";
+import { ViewerView } from "./room/ViewerView";
 
 interface Props {
   text: string;
@@ -66,12 +36,7 @@ export default function Content({
   role,
 }: Props) {
   const textareaRef = useRef<HTMLTextAreaElement>(null);
-  const [recording, setRecording] = useState(false);
   const [transcriptionMethod, setTranscriptionMethod] = useState<TranscriptionMethod>("server");
-  const [isPushToTalkActive, setIsPushToTalkActive] = useState(false);
-  const recordingManagerRef = useRef<RecordingManager | null>(null);
-  const { transcript, resetTranscript, browserSupportsSpeechRecognition } =
-    useSpeechRecognition();
 
   // Tldraw state
   const [inputMode, setInputMode] = useState<"text" | "drawing">("text");
@@ -85,231 +50,6 @@ export default function Content({
   const editorRef = useRef<any>(null);
   const lastEmittedSnapshotRef = useRef<string>("");
 
-  // Initialize recording manager (client-side only)
-  useEffect(() => {
-    if (typeof window !== "undefined") {
-      recordingManagerRef.current = new RecordingManager({
-        method: transcriptionMethod,
-        onChunkTranscribed: (transcribedText) => {
-          if (transcribedText.trim()) {
-            setText((prev) => {
-              const newText = prev + " " + transcribedText;
-              debouncedEmitText(newText);
-              return newText;
-            });
-          }
-        },
-        onError: (error) => {
-          toast.error(error);
-        },
-        isRecording: recording,
-      });
-    }
-  }, [transcriptionMethod, setText]);
-
-  // Tldraw persistence and socket handling
-  useLayoutEffect(() => {
-    setDrawingLoadingState({ status: "loading" });
-
-    // Listen for drawing data from socket
-    if (socket) {
-      socket.on("receive-drawing", (data: string, username: string) => {
-        if (username === getCookie("username")) return;
-        if (data) {
-          try {
-            // Set flag to prevent emitting back to server
-            isLoadingFromSocket.current = true;
-            const snapshot = JSON.parse(data);
-            loadSnapshot(store, snapshot);
-            // Update last emitted snapshot so we don't re-emit what we just received
-            lastEmittedSnapshotRef.current = data;
-            // Reset flag after a short delay to ensure store listener doesn't fire
-            setTimeout(() => {
-              isLoadingFromSocket.current = false;
-            }, 200);
-          } catch (error: any) {
-            console.error("Error loading drawing:", error);
-            setDrawingLoadingState({ status: "error", error: error.message });
-            isLoadingFromSocket.current = false;
-          }
-        }
-      });
-    }
-
-    setDrawingLoadingState({ status: "ready" });
-
-    // Setup store listener to emit changes to server with debounce
-    const debouncedEmitDrawing = debounce(async () => {
-      // Don't emit if we're currently loading from socket
-      if (isLoadingFromSocket.current) {
-        console.log("Skipping emit - loading from socket");
-        return;
-      }
-
-      const snapshot = getSnapshot(store);
-      const snapshotString = JSON.stringify(snapshot);
-
-      // Only emit if snapshot actually changed
-      if (snapshotString === lastEmittedSnapshotRef.current) {
-        console.log("Skipping emit - no changes");
-        return;
-      }
-
-      if (socket && socket.connected) {
-        console.log("Sending drawing");
-        // Export drawing as image for LaTeX conversion
-        const imageData = await exportDrawingAsImage();
-
-        // Send both drawing data and image data in one event
-        socket.emit("send-drawing", snapshotString, router.asPath.split("#")[1], null, imageData);
-
-        // Update last emitted snapshot
-        lastEmittedSnapshotRef.current = snapshotString;
-      }
-    }, 1000);
-
-    const cleanupFn = store.listen(() => {
-      debouncedEmitDrawing();
-    });
-
-    // Function to export drawing as image
-    const exportDrawingAsImage = async () => {
-      try {
-        if (!editorRef.current) {
-          return null;
-        }
-
-        const editor = editorRef.current;
-        const shapeIds = editor.getCurrentPageShapeIds();
-
-        if (shapeIds.size === 0) {
-          // No shapes to export, skip
-          return null;
-        }
-
-        const blob = await exportToBlob({
-          editor,
-          ids: Array.from(shapeIds),
-          format: "png",
-          opts: { background: true, padding: 16, scale: 2 },
-        });
-
-        // Convert blob to base64
-        return new Promise((resolve) => {
-          const reader = new FileReader();
-          reader.readAsDataURL(blob);
-          reader.onloadend = () => {
-            resolve(reader.result);
-          };
-        });
-      } catch (error) {
-        console.error("Error exporting drawing:", error);
-        return null;
-      }
-    };
-
-    return () => {
-      cleanupFn();
-      if (socket) {
-        socket.off("receive-drawing");
-      }
-    };
-  }, [store, socket, router]);
-
-  // Handle browser speech recognition for browser method
-  useEffect(() => {
-    if (transcript && transcriptionMethod === "browser") {
-      setText((prev) => {
-        const newText = prev + " " + transcript;
-        debouncedEmitText(newText);
-        return newText;
-      });
-      resetTranscript();
-    }
-  }, [transcript, setText, resetTranscript, transcriptionMethod]);
-
-  const startRecording = async () => {
-    if (typeof window === "undefined") {
-      toast.error("Recording not available on server side");
-      return;
-    }
-
-    if (transcriptionMethod === "browser") {
-      if (!browserSupportsSpeechRecognition) {
-        toast.error("Your browser doesn't support speech recognition");
-        return;
-      }
-      setRecording(true);
-      SpeechRecognition.startListening({ continuous: true });
-      toast.success("Browser voice recording started");
-    } else {
-      // Server method using MediaRecorder
-      if (!recordingManagerRef.current) {
-        toast.error("Recording manager not initialized - please wait");
-        return;
-      }
-      try {
-        await recordingManagerRef.current?.startRecording();
-        setRecording(true);
-        toast.success("Server voice recording started");
-      } catch (error) {
-        toast.error("Failed to start recording");
-      }
-    }
-  };
-
-  const stopRecording = () => {
-    if (transcriptionMethod === "browser") {
-      setRecording(false);
-      SpeechRecognition.stopListening();
-      resetTranscript();
-      toast.success("Browser voice recording stopped");
-    } else {
-      // Server method
-      recordingManagerRef.current?.stopRecording();
-      setRecording(false);
-      toast.success("Server voice recording stopped");
-    }
-  };
-
-  // Push-to-talk: hold Ctrl (Windows/Linux) or Meta/Command (Mac)
-  useEffect(() => {
-    if (!recordingManagerRef.current) return;
-    const handleKeyDown = async (e: KeyboardEvent) => {
-
-      // Only respond to Ctrl key (Windows/Linux) or Command key (Mac)
-      // Ignore other modifier keys like Windows key, Alt, etc.
-      const isCtrlKey = e.key === 'Control' || e.code === 'ControlLeft' || e.code === 'ControlRight';
-
-      if ((!isCtrlKey) || isPushToTalkActive) return;
-
-      // Start PTT only if not already recording (avoid conflicts)
-      try {
-        await recordingManagerRef.current?.startPushToTalk();
-        setIsPushToTalkActive(true);
-      } catch { }
-    };
-    const handleKeyUp = async (e: KeyboardEvent) => {
-
-      // Only respond to Ctrl key (Windows/Linux) or Command key (Mac)
-      const isCtrlKey = e.key === 'Control' || e.code === 'ControlLeft' || e.code === 'ControlRight';
-
-      if ((!isCtrlKey) || !isPushToTalkActive) return;
-
-      // Key released -> stop PTT
-      try {
-        await recordingManagerRef.current?.stopPushToTalk();
-        setIsPushToTalkActive(false);
-      } catch { }
-    };
-    window.addEventListener("keydown", handleKeyDown);
-    window.addEventListener("keyup", handleKeyUp);
-    return () => {
-      window.removeEventListener("keydown", handleKeyDown);
-      window.removeEventListener("keyup", handleKeyUp);
-    };
-  }, [isPushToTalkActive]);
-
   // Debounced socket emission using utility function
   const debouncedEmitText = useMemo(
     () =>
@@ -320,6 +60,57 @@ export default function Content({
       }, 800),
     [socket, router],
   );
+
+  // Voice recording hook
+  const { recording, isPushToTalkActive, startRecording, stopRecording } =
+    useVoiceRecording({
+      transcriptionMethod,
+      setText,
+      debouncedEmitText,
+    });
+
+  // Tldraw persistence and socket handling
+  useLayoutEffect(() => {
+    setDrawingLoadingState({ status: "loading" });
+
+    // Listen for drawing data from socket
+    if (socket) {
+      socket.on("receive-drawing", (data: string, username: string) => {
+        handleReceiveDrawing(
+          data,
+          username,
+          store,
+          isLoadingFromSocket,
+          lastEmittedSnapshotRef,
+          setDrawingLoadingState
+        );
+      });
+    }
+
+    setDrawingLoadingState({ status: "ready" });
+
+    // Setup store listener to emit changes to server with debounce
+    const roomId = router.asPath.split("#")[1];
+    const debouncedEmitDrawing = createDrawingEmitter(
+      store,
+      socket,
+      roomId,
+      isLoadingFromSocket,
+      lastEmittedSnapshotRef,
+      editorRef
+    );
+
+    const cleanupFn = store.listen(() => {
+      debouncedEmitDrawing();
+    });
+
+    return () => {
+      cleanupFn();
+      if (socket) {
+        socket.off("receive-drawing");
+      }
+    };
+  }, [store, socket, router]);
 
   const handleCopyLatex = () => {
     copyToClipboard(latex, "LaTeX code copied to clipboard");
@@ -354,259 +145,38 @@ export default function Content({
     // For viewers, keep the tabbed interface
     if (role !== "editor") {
       return (
-        <Tabs defaultValue="preview" className="flex-1 flex flex-col">
-          <motion.div variants={itemVariants}>
-            <TabsList className="grid w-full grid-cols-1 mb-6">
-              <TabsTrigger value="preview" className="flex items-center gap-2">
-                <BookOpen className="h-4 w-4" />
-                Preview
-              </TabsTrigger>
-            </TabsList>
-          </motion.div>
-
-          <TabsContent
-            value="preview"
-            className="flex-1 flex flex-col space-y-4"
-          >
-            <motion.div variants={itemVariants} className="flex-1">
-              <Card className="h-full flex flex-col">
-                <CardHeader className="pb-2">
-                  <CardTitle>LaTeX Preview</CardTitle>
-                  <CardDescription>
-                    Rendered output of LaTeX equations
-                  </CardDescription>
-                </CardHeader>
-                <CardContent className="flex-1 flex flex-col">
-                  <div className="flex-1 bg-white p-6 rounded-md border overflow-auto">
-                    <LatexDisplayer latex={latex} />
-                  </div>
-
-                  <div className="flex flex-wrap gap-2 mt-4">
-                    <motion.div
-                      whileHover={{ scale: 1.05 }}
-                      whileTap={{ scale: 0.95 }}
-                    >
-                      <Button
-                        onClick={handleCopyLatex}
-                        variant="outline"
-                        className="flex items-center gap-2"
-                      >
-                        <Copy className="h-4 w-4" />
-                        Copy LaTeX
-                      </Button>
-                    </motion.div>
-
-                    <motion.div
-                      whileHover={{ scale: 1.05 }}
-                      whileTap={{ scale: 0.95 }}
-                    >
-                      <Button
-                        onClick={handleDownloadPDF}
-                        variant="outline"
-                        className="flex items-center gap-2"
-                      >
-                        <Download className="h-4 w-4" />
-                        Download PDF
-                      </Button>
-                    </motion.div>
-                  </div>
-                </CardContent>
-              </Card>
-            </motion.div>
-          </TabsContent>
-        </Tabs>
+        <ViewerView
+          latex={latex}
+          onCopyLatex={handleCopyLatex}
+          onDownloadPDF={handleDownloadPDF}
+          itemVariants={itemVariants}
+        />
       );
     }
 
     // For editors, show split view with editor and preview side by side
     return (
-      <div className="flex-1 flex flex-col md:flex-row gap-1">
-        <motion.div variants={itemVariants} className="flex-1 min-w-0">
-          <Card className="h-full flex flex-col">
-            <CardHeader className="pb-2">
-              <div className="flex items-center justify-between">
-                <div>
-                  <CardTitle>
-                    {inputMode === "text" ? "LaTeX Editor" : "Drawing Board"}
-                  </CardTitle>
-                  <CardDescription>
-                    {inputMode === "text"
-                      ? "Type or use voice input to create LaTeX equations"
-                      : "Draw diagrams and illustrations"}
-                  </CardDescription>
-                </div>
-                <div className="flex gap-2">
-                  <Button
-                    onClick={() => setInputMode("text")}
-                    variant={inputMode === "text" ? "default" : "outline"}
-                    size="sm"
-                    className="flex items-center gap-2"
-                  >
-                    <Type className="h-4 w-4" />
-                    Text
-                  </Button>
-                  <Button
-                    onClick={() => setInputMode("drawing")}
-                    variant={inputMode === "drawing" ? "default" : "outline"}
-                    size="sm"
-                    className="flex items-center gap-2"
-                  >
-                    <Pencil className="h-4 w-4" />
-                    Draw
-                  </Button>
-                </div>
-              </div>
-            </CardHeader>
-            <CardContent className="flex-1 flex flex-col">
-              {inputMode === "text" ? (
-                <>
-                  <Textarea
-                    ref={textareaRef}
-                    className="w-full flex-1 text-lg resize-none font-mono"
-                    value={text}
-                    onChange={(e) => {
-                      const newText = e.target.value;
-                      setText(newText);
-                      debouncedEmitText(newText);
-                    }}
-                    placeholder="Start typing here or use voice input..."
-                  />
-
-                  <div className="flex flex-col gap-3 mt-4">
-                    <motion.div
-                      className={`flex items-center justify-center gap-3 px-6 py-4 rounded-lg text-base font-medium transition-all duration-200 ${isPushToTalkActive
-                        ? "bg-gradient-to-r from-red-500 to-pink-500 text-white shadow-lg scale-105"
-                        : "bg-gradient-to-r from-purple-500 to-pink-500 text-white shadow-md hover:shadow-lg hover:scale-[1.02]"
-                        }`}
-                      whileHover={{ scale: isPushToTalkActive ? 1.05 : 1.02 }}
-                      whileTap={{ scale: 0.98 }}
-                    >
-                      <Mic className={`h-5 w-5 ${isPushToTalkActive ? "animate-pulse" : ""}`} />
-                      <span>
-                        {isPushToTalkActive ? (
-                          <>
-                            <span className="font-bold">🎙️ Recording...</span> Release{" "}
-                            <kbd className="px-2 py-1 bg-white/20 border border-white/30 rounded text-sm font-mono ml-1">
-                              Ctrl
-                            </kbd>
-                          </>
-                        ) : (
-                          <>
-                            <span className="font-semibold">Press & Hold</span>{" "}
-                            <kbd className="px-2 py-1 bg-white/20 border border-white/30 rounded text-sm font-mono mx-1">
-                              Ctrl
-                            </kbd>{" "}
-                            <span className="font-semibold">to Speak</span>
-                          </>
-                        )}
-                      </span>
-                    </motion.div>
-
-                    <div className="flex items-center gap-2">
-                      <div className="flex-1 border-t border-gray-300"></div>
-                      <span className="text-xs text-gray-500 uppercase tracking-wide">or</span>
-                      <div className="flex-1 border-t border-gray-300"></div>
-                    </div>
-
-                    <motion.div
-                      whileHover={{ scale: 1.02 }}
-                      whileTap={{ scale: 0.98 }}
-                    >
-                      <Button
-                        onClick={recording ? stopRecording : startRecording}
-                        variant={recording ? "destructive" : "outline"}
-                        className="w-full flex items-center justify-center gap-2"
-                      >
-                        {recording ? (
-                          <>
-                            <MicOff className="h-4 w-4" />
-                            Stop Continuous Recording
-                          </>
-                        ) : (
-                          <>
-                            <Mic className="h-4 w-4" />
-                            Start Continuous Voice Input
-                          </>
-                        )}
-                      </Button>
-                    </motion.div>
-                  </div>
-                </>
-              ) : (
-                <div className="flex-1 w-full h-full min-h-[500px]">
-                  {drawingLoadingState.status === "loading" && (
-                    <div className="flex items-center justify-center h-full">
-                      <DefaultSpinner />
-                    </div>
-                  )}
-                  {drawingLoadingState.status === "error" && (
-                    <div className="flex flex-col items-center justify-center h-full">
-                      <h2 className="text-xl font-bold text-red-600">Error!</h2>
-                      <p className="text-gray-600">{drawingLoadingState.error}</p>
-                    </div>
-                  )}
-                  {drawingLoadingState.status === "ready" && (
-                    <div className="w-full h-full">
-                      <Tldraw
-                        store={store}
-                        onMount={(editor) => {
-                          editorRef.current = editor;
-                        }}
-                      />
-                    </div>
-                  )}
-                </div>
-              )}
-            </CardContent>
-          </Card>
-        </motion.div>
-
-        <motion.div variants={itemVariants} className="flex-1 min-w-0">
-          <Card className="h-full flex flex-col">
-            <CardHeader className="pb-2">
-              <CardTitle>LaTeX Preview</CardTitle>
-              <CardDescription>
-                Rendered output of your LaTeX equations
-              </CardDescription>
-            </CardHeader>
-            <CardContent className="flex-1 flex flex-col">
-              <div className="flex-1 bg-white p-6 rounded-md border overflow-auto">
-                <LatexDisplayer latex={latex} />
-              </div>
-
-              <div className="flex flex-wrap gap-2 mt-4">
-                <motion.div
-                  whileHover={{ scale: 1.05 }}
-                  whileTap={{ scale: 0.95 }}
-                >
-                  <Button
-                    onClick={handleCopyLatex}
-                    variant="outline"
-                    className="flex items-center gap-2"
-                  >
-                    <Copy className="h-4 w-4" />
-                    Copy LaTeX
-                  </Button>
-                </motion.div>
-
-                <motion.div
-                  whileHover={{ scale: 1.05 }}
-                  whileTap={{ scale: 0.95 }}
-                >
-                  <Button
-                    onClick={handleDownloadPDF}
-                    variant="outline"
-                    className="flex items-center gap-2"
-                  >
-                    <Download className="h-4 w-4" />
-                    Download PDF
-                  </Button>
-                </motion.div>
-              </div>
-            </CardContent>
-          </Card>
-        </motion.div>
-      </div>
+      <EditorView
+        inputMode={inputMode}
+        onInputModeChange={setInputMode}
+        text={text}
+        setText={setText}
+        textareaRef={textareaRef}
+        debouncedEmitText={debouncedEmitText}
+        isPushToTalkActive={isPushToTalkActive}
+        recording={recording}
+        onStartRecording={startRecording}
+        onStopRecording={stopRecording}
+        drawingLoadingState={drawingLoadingState}
+        store={store}
+        onDrawingMount={(editor) => {
+          editorRef.current = editor;
+        }}
+        latex={latex}
+        onCopyLatex={handleCopyLatex}
+        onDownloadPDF={handleDownloadPDF}
+        itemVariants={itemVariants}
+      />
     );
   };
